:recycle: refactor: extract typing
