+++ jobs/Frontend/mews-movie-search-app/src/useMovies.tsx
@@ -1,33 +1,11 @@
 import React from "react";
+import { MovieListResponse } from "./types/movies";
+import { domainURL } from "./utils/constant";
 
-export type MovieListResponse = {
-  page: number;
-  results: Movie[];
-  total_pages: number;
-  total_results: number;
-};
-
-export type Movie = {
-  adult: boolean;
-  backdrop_path: string;
-  genre_ids: number[];
-  id: number;
-  original_language: string;
-  original_title: string;
-  overview: string;
-  popularity: number;
-  poster_path: string;
-  release_date: string;
-  title: string;
-  video: boolean;
-  vote_average: number;
-  vote_count: number;
-};
 type useMoviesArgTypes = {
   page?: number;
   searchTerm?: string;
 };
-const domainURL = `https://api.themoviedb.org/3/`;
 const emptyMovieListResponse: MovieListResponse = {
   results: [],
   page: 0,
